+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 - Allow to inject a promotion hook to ExpectTests instead of a static boolean. This allows to, e.g. prompt the developer with a diff before promoting, or saving diff to files along test run.
+- [Diff utilities](src/main/kotlin/nyub/expekt/diff) and [PromptWithPanel](src/main/kotlin/nyub/expekt/diff/PromptWithDiffPanel.kt) promotion trigger that displays a pop-up prompting to promote or reject a snapshot
 - Allow to trigger promotion with a `promote@` label before expected string block (Kotlin only)
 - Check or promote all expect calls within a test before raising assertion errors, with soft assertions
 ```kotlin